Extract a Vue dashboard component.
